server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# LangChain / LangGraph
langchain-core==1.2.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-sdk==0.3.0
langsmith==0.4.59
langchain-google-genai==4.0.0
langchain-tavily>=0.2.14

tiktoken==0.5.2

# Pydantic
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
typing-extensions==4.15.0

# Gemini SDK
google-generativeai==0.8.3
google-genai==1.55.0

# HTTP
httpx==0.28.1
aiohttp>=3.11.14,<4.0.0

# Search
duckduckgo-search==6.4.2
tavily-python>=0.5.0

# Utils
tenacity==9.1.2
packaging==25.0
PyYAML==6.0.3
jsonpatch==1.33
